// src/react-app-env.d.ts
/// <reference types="react-scripts" />

export interface Apparel {
    freeEntry: CheckboxData;
    big: CheckboxData;
    medium: CheckboxData;
    small: CheckboxData;
    height: CheckboxData;
    sleeve: CheckboxData;
    neck: CheckboxData;
    shape: CheckboxData;
    design: CheckboxData;
    material: CheckboxData;
    set: CheckboxData;
    image: CheckboxData;
    pattern: CheckboxData;
    seasonOrEvent: CheckboxData;
    ageOrSex: CheckboxData;
    place: CheckboxData;
    others: CheckboxData;
    size: CheckboxData;
    color: CheckboxData;
}

export interface Shoes {
    freeEntry: CheckboxData;
    big: CheckboxData;
    medium: CheckboxData;
    heel: CheckboxData;
    designOrShape: CheckboxData;
    material: CheckboxData;
    image: CheckboxData;
    pattern: CheckboxData;
    seasonOrEvent: CheckboxData;
    ageOrSex: CheckboxData;
    place: CheckboxData;
    others: CheckboxData;
    size: CheckboxData;
    color: CheckboxData;
}

export interface BagWallet {
    freeEntry: CheckboxData;
    big: CheckboxData;
    mediumBag: CheckboxData;
    mediumWallet: CheckboxData;
    functionBag: CheckboxData;
    functionWallet: CheckboxData;
    designOrShape: CheckboxData;
    material: CheckboxData;
    image: CheckboxData;
    pattern: CheckboxData;
    ageOrSex: CheckboxData;
    place: CheckboxData;
    others: CheckboxData;
    seasonOrEvent: CheckboxData;
    size: CheckboxData;
    color: CheckboxData;
}

export interface Dress {
    freeEntry: CheckboxData;
    big: CheckboxData;
    medium: CheckboxData;
    height: CheckboxData;
    sleeve: CheckboxData;
    neck: CheckboxData;
    shape: CheckboxData;
    design: CheckboxData;
    material1: CheckboxData;
    material2: CheckboxData;
    image: CheckboxData;
    pattern1: CheckboxData;
    pattern2: CheckboxData;
    ageOrSex: CheckboxData;
    place: CheckboxData;
    others: CheckboxData;
    seasonOrEvent: CheckboxData;
    size: CheckboxData;
    color: CheckboxData;
}

export interface Swimsuit {
    freeEntry: CheckboxData;
    big: CheckboxData;
    medium: CheckboxData;
    shape: CheckboxData;
    design: CheckboxData;
    set: CheckboxData;
    image: CheckboxData;
    pattern: CheckboxData;
    ageOrSex: CheckboxData;
    others: CheckboxData;
    size: CheckboxData;
    color: CheckboxData;
}